Pull up documentation property to RustItem.
