+++ src/main/kotlin/org/rust/lang/core/psi/RustNamedElement.kt
@@ -6,6 +6,5 @@ import com.intellij.psi.PsiNamedElement
 interface RustNamedElement   : RustCompositeElement
                              , PsiNamedElement
                              , NavigatablePsiElement {
-    val documentation: String?
 }
